Update 1472G.cpp
Updated 1472G as it got TLE on sys test cases (test 10). Fixed dumb issue of not checking when going through a node (thought it wasn't necessary as it was a DAG, but of course given the case many smaller nodes go to some other is possible, making the dfs go through it, and its children, several times).